♻️ refactor: Extract update checking logic into separate module
Moves update detection and processing functionality from main extension file into dedicated updates module.

• Separates update checking concerns from activation logic
• Adds comprehensive documentation for update handling functions
